Building the project's architecture, installing dependencies, and applying global styles
Building the setup of the React project 'Rocket Movies' with Vite. Removing the files that won't be used, installing the project's dependencies, which are the libraries 'styled-components,' 'react-router-dom,' and 'react-icons,' and structuring the project architecture by creating the 'components' folder, the 'pages' folder, the 'routes' folder, and the 'styles' folder. Finally, applying the global styles of the application to the 'main.jsx' file and creating the 'Routes' component that will manage the application's routes, contained within the 'AppRoutes' and 'AuthRoutes' components