Building the 'Home' page component by creating the necessary components
Creating the 'Header,' 'Input,' 'Button,' 'IconButton,' 'Tag,' 'Stars,' 'Note,' and 'title' components, each with their respective 'index.jsx' files (containing the component structure) and 'styles.js' files (containing the component styles). Updating the global styles file 'styles.js' and the 'theme.js' file that contains the theme data for the application. Finally, creating, building the structure, and applying styles to the component of the home page, the 'Home' component, using the other components created to compose the page and updating the 'app.routes.jsx' file to render the 'Home' page on the '/' route