Creating the 'New' page and the components used in it
Creating the components 'TextArea', 'LinkToBack', and 'NoteMarker' to compose the structure of the 'New' page. Updating the global style files, 'reset.css' and 'normalize.css,' and updating the 'Input,' 'IconButton,' and 'Stars' components. Finally, building the structure, applying styles, and creating the component for the new note page, the 'New' component, and updating the 'app.routes.jsx' file to render the new page at the '/new' route