Creating the components for the 'Datails' and 'Profile' pages
Updating the structure of the components 'Input,' 'Note,' 'LinkToBack,' 'Stars,' and 'Title' to use them in building the new pages. Updating the styles and content of the 'Home' and 'New' pages. Building two more components, the 'Details' component, which contains detailed information about a note written by the user, and the 'Profile' component, which is the user's profile page. Finally, adding the routes for the created pages in the 'app.routes.jsx' file