Creating the 'SignIn' (login) and 'SignUp' (registration) pages.
Updating the 'Home', 'New', and 'Details' pages by standardizing the styles, spacing, sizes, and content on the pages. Updating the 'LinkToBack,' 'Note,' and 'Header' components. Finally, creating the components for the user login pages, the 'SignIn' component, and for the user registration page, the 'SignUp' component, and completing the construction of the application's pages by updating the application routes.