Initiating the integration with the Back-End, installing new dependencies, and building the logics for the 'SignUp' and 'SignIn' pages
Starting the final stage of the application development, initiating the integration with the Back-End. Installing new dependencies such as the 'axios' library for communicating with the Back-End API through HTTP requests and 'sweetalert2' for creating custom alerts and pop-ups according to the application's theme. Creating the 'services' folder and configuring the 'axios' library to make requests to the Back-End. Establishing the 'validations' folder to store files with validation logic, like the 'form' file that provides email and password validations through regular expressions. Creating the 'utils' folder with the 'alert' folder and the 'alert.jsx' file, containing functions that trigger custom alerts created with the help of the 'sweetalert2' library. Implementing the logic to register users with the 'SignUp' component, validating the fields, storing user-provided data in states, and sending this data to the API. Finally, creating the 'contexts' and 'hooks' folders, establishing the authentication context 'AuthContext,' creating the 'useAuth' hook, and implementing the login logic in the 'SignIn' component