Creating the logic to update user profile data through the 'Profile' component
Continuing the integration with the Back-End. Updating the authentication context by adding the 'updateProfile' function, which contains the logic responsible for updating user data. It takes an object as an argument containing the new data that will be sent to the API to update the user data in the database. Modifying the 'Header' component to fetch the user's avatar image from the Back-End. Importing the 'avatar_placeholder.svg' image to serve as a default avatar image for all users. Finally, implementing the logic to update the user through the 'Profile' component. It retrieves user data from the form fields, applies validations to the fields and received data, and implements the 'updateProfile' function from the authentication context to perform the data update